refactor: move all RequirementAgent files to the same folder
