+++ python/beeai_framework/agents/experimental/requirements/_utils.py
@@ -16,6 +16,18 @@ from typing import Any, TypeVar
 
 from beeai_framework.errors import FrameworkError
 from beeai_framework.tools import AnyTool, Tool
+from beeai_framework.agents.experimental.requirements.prompts import (
+    RequirementAgentSystemPromptInput,
+    RequirementAgentToolTemplateDefinition,
+)
+from beeai_framework.agents.experimental.requirements.types import RequirementAgentRequest
+from beeai_framework.backend import SystemMessage
+from beeai_framework.template import PromptTemplate
+from beeai_framework.utils.strings import to_json
+
+
+from beeai_framework.tools import AnyTool
+from beeai_framework.tools.tool import Tool
 
 T = TypeVar("T", bound=Any)
 
@@ -92,3 +104,23 @@ def _extract_target_name(target: TargetType) -> str:
         return target.name
     else:
         return target.__qualname__
+
+
+def _create_system_message(
+    *, template: PromptTemplate[RequirementAgentSystemPromptInput], request: RequirementAgentRequest
+) -> SystemMessage:
+    return SystemMessage(
+        template.render(
+            tools=[
+                RequirementAgentToolTemplateDefinition.from_tool(tool, allowed=tool in request.allowed_tools)
+                for tool in request.tools
+            ],
+            final_answer_name=request.final_answer.name,
+            final_answer_schema=to_json(
+                request.final_answer.input_schema.model_json_schema(mode="validation"), indent=2
+            )
+            if request.final_answer.custom_schema
+            else None,
+            final_answer_instructions=request.final_answer.instructions,
+        )
+    )
