+++ python/examples/agents/experimental/requirement/complex.py
@@ -2,7 +2,7 @@ import asyncio
 import math
 
 from beeai_framework.agents.experimental import RequirementAgent
-from beeai_framework.agents.experimental.prompts import (
+from beeai_framework.agents.experimental.requirements.prompts import (
     RequirementAgentCycleDetectionPrompt,
     RequirementAgentSystemPrompt,
     RequirementAgentTaskPrompt,
@@ -12,7 +12,7 @@ from beeai_framework.agents.experimental.prompts import (
 from beeai_framework.agents.experimental.requirements import Requirement, Rule
 from beeai_framework.agents.experimental.requirements.conditional import ConditionalRequirement
 from beeai_framework.agents.experimental.requirements.requirement import run_with_context
-from beeai_framework.agents.experimental.types import RequirementAgentRunState, RequirementAgentTemplates
+from beeai_framework.agents.experimental.requirements.types import RequirementAgentRunState, RequirementAgentTemplates
 from beeai_framework.backend import ChatModel
 from beeai_framework.context import RunContext
 from beeai_framework.memory import UnconstrainedMemory
