+++ python/examples/agents/experimental/requirement/custom_requirement.py
@@ -3,7 +3,7 @@ import asyncio
 from beeai_framework.agents.experimental import RequirementAgent
 from beeai_framework.agents.experimental.requirements import Requirement, Rule
 from beeai_framework.agents.experimental.requirements.requirement import run_with_context
-from beeai_framework.agents.experimental.types import RequirementAgentRunState
+from beeai_framework.agents.experimental.requirements.types import RequirementAgentRunState
 from beeai_framework.backend import AssistantMessage, ChatModel
 from beeai_framework.context import RunContext
 from beeai_framework.middleware.trajectory import GlobalTrajectoryMiddleware
